# 参考及文档说明/接口文档/智能工单系统 API.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "智能工单系统 API",
    "description": "智能工单管理系统接口文档",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:10001",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/role/{roleId}/permissions": {
      "get": {
        "tags": [
          "角色管理"
        ],
        "summary": "角色已分配权限ID列表",
        "operationId": "listPermissions",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListLong"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "角色管理"
        ],
        "summary": "分配权限",
        "operationId": "assignPermissions",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignPermissionDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/department/{id}": {
      "put": {
        "tags": [
          "部门管理"
        ],
        "summary": "修改部门",
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "部门管理"
        ],
        "summary": "删除部门",
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/department": {
      "post": {
        "tags": [
          "部门管理"
        ],
        "summary": "新增部门",
        "operationId": "add",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResultVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {},
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "AssignPermissionDTO": {
        "type": "object",
        "description": "分配权限请求",
        "properties": {
          "permissionIds": {
            "type": "array",
            "description": "权限ID列表,可为空列表表示清空",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        },
        "required": [
          "permissionIds"
        ]
      },
      "DepartmentDTO": {
        "type": "object",
        "description": "部门新增/修改请求参数",
        "properties": {
          "deptCode": {
            "type": "string",
            "description": "部门编码(唯一,代码判断用)",
            "maxLength": 20,
            "minLength": 0
          },
          "deptName": {
            "type": "string",
            "description": "部门名称",
            "maxLength": 50,
            "minLength": 0
          },
          "remark": {
            "type": "string",
            "description": "备注,可空",
            "maxLength": 200,
            "minLength": 0
          }
        },
        "required": [
          "deptCode",
          "deptName"
        ]
      },
...
      "ResultListLong": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    }
  }
}